requirements.txt
python-firebase==1.2.1